PF: - Reorder My structure
