+++ My/templates/My/Fridge/fridge_view.html
@@ -6,7 +6,7 @@
     <div class="row">
       <div class="col-md-12 blog-main">
         <div class="blog-post">
-          {% include 'My/fridges_template.html' %}
+          {% include 'My/Fridge/fridges_template.html' %}
 
           <div>
             {% for product in products %}
