+++ My/templates/My/ShoppingList/shopping_list_product_add.html
