+++ My/templates/My/home_index.html
@@ -5,7 +5,7 @@
   <main role="main" class="container">
     <div class="row">
       <div class="col-md-12 blog-main">
-        {% include 'My/fridges_template.html' %}
+        {% include 'My/Fridge/fridges_template.html' %}
       </div>
     </div>
   </main>
